ml-models/requirements.txt
PyPDF2==3.0.1
python-docx==1.0.1
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6